180/360 support for both video and photo domes
